Refactor outputter
