+++ cmd/flags/update_test.go
@@ -80,7 +80,7 @@ func TestUpdate(t *testing.T) {
 		client := flags.MockClient{}
 		client.
 			On("Update", mockArgs...).
-			Return([]byte(`{}`), errors.NewError("An error"))
+			Return([]byte(`{}`), errors.NewError(`{"message": "An error"}`))
 		clients := cmd.APIClients{
 			FlagsClient: &client,
 		}
@@ -207,7 +207,7 @@ func TestToggle(t *testing.T) {
 		client := flags.MockClient{}
 		client.
 			On("Update", mockArgs...).
-			Return([]byte(`{}`), errors.NewError("An error"))
+			Return([]byte(`{}`), errors.NewError(`{"message": "An error"}`))
 		clients := cmd.APIClients{
 			FlagsClient: &client,
 		}
