+++ internal/output/output_test.go
@@ -1,1 +1,3 @@
 package output_test
+
+//
