+++ internal/output/singular_outputter_test.go
@@ -1,41 +1,32 @@
 package output_test
 
-import (
-	"testing"
+// func TestSingularOutputter_JSON(t *testing.T) {
+// 	input := []byte(`{
+// 		"key": "test-key",
+// 		"name": "test-name",
+// 		"other": "another-value"
+// 	}`)
+// 	output, err := output.CmdOutput(
+// 		"json",
+// 		output.NewSingularOutput(input),
+// 	)
 
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
+// 	require.NoError(t, err)
+// 	assert.JSONEq(t, output, string(input))
+// }
 
-	"ldcli/internal/output"
-)
+// func TestSingularOutputter_String(t *testing.T) {
+// 	input := []byte(`{
+// 		"key": "test-key",
+// 		"name": "test-name",
+// 		"other": "another-value"
+// 	}`)
+// 	expected := "test-name (test-key)"
+// 	output, err := output.CmdOutput(
+// 		"plaintext",
+// 		output.NewSingularOutput(input),
+// 	)
 
-func TestSingularOutputter_JSON(t *testing.T) {
-	input := []byte(`{
-		"key": "test-key",
-		"name": "test-name",
-		"other": "another-value"
-	}`)
-	output, err := output.CmdOutput(
-		"json",
-		output.NewSingularOutput(input),
-	)
-
-	require.NoError(t, err)
-	assert.JSONEq(t, output, string(input))
-}
-
-func TestSingularOutputter_String(t *testing.T) {
-	input := []byte(`{
-		"key": "test-key",
-		"name": "test-name",
-		"other": "another-value"
-	}`)
-	expected := "test-name (test-key)"
-	output, err := output.CmdOutput(
-		"plaintext",
-		output.NewSingularOutput(input),
-	)
-
-	require.NoError(t, err)
-	assert.Equal(t, expected, output)
-}
+// 	require.NoError(t, err)
+// 	assert.Equal(t, expected, output)
+// }
